[COOK-4434] writing a timestamp: not convergent
Every time Chef runs, it will update the template. This is totally
idempotent, however it is not convergent. In a world where report
handlers will track the number of resources updated in a Chef run, and
potentially the timestamp of that run (depending on implmentation),
having the last run's time stamp in the motd is less useful.